# artifacts/api-server/src/routes/index.ts
import { Router, type IRouter, type Request, type Response, type NextFunction } from "express";
import healthRouter from "./health";
import authRouter from "./auth";
import columnsRouter from "./columns";
import tasksRouter from "./tasks";

const router: IRouter = Router();

export function requireAuth(req: Request, res: Response, next: NextFunction) {
  if (!req.isAuthenticated()) {
    res.status(401).json({ error: "Unauthorized" });
    return;
  }
  next();
}

export function requireAdmin(req: Request, res: Response, next: NextFunction) {
  if (!req.isAuthenticated()) {
    res.status(401).json({ error: "Unauthorized" });
    return;
  }
  if (req.user.role !== "admin") {
    res.status(403).json({ error: "Forbidden: admin access required" });
    return;
  }
  next();
}

router.use(healthRouter);
router.use(authRouter);
router.use("/columns", requireAuth, columnsRouter);
router.use("/tasks", requireAuth, tasksRouter);

export default router;