# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nature Kanban Task Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AuthUserRole = (typeof AuthUserRole)[keyof typeof AuthUserRole];

export const AuthUserRole = {
  admin: "admin",
  user: "user",
} as const;

export interface AuthUser {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
  role: AuthUserRole;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  state: string;
  nonce?: string | null;
  code_verifier: string;
  redirect_uri: string;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export interface Column {
  id: number;
  title: string;
  position: number;
  createdAt: string;
}

export interface CreateColumnRequest {
  title: string;
  position: number;
}

export interface UpdateColumnRequest {
  title?: string;
  position?: number;
}

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Task {
  id: number;
  columnId: number;
  userId?: string | null;
  title: string;
  description?: string | null;
  priority: TaskPriority;
  position: number;
  createdAt: string;
}

export type CreateTaskRequestPriority =
  (typeof CreateTaskRequestPriority)[keyof typeof CreateTaskRequestPriority];

export const CreateTaskRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateTaskRequest {
  columnId: number;
  title: string;
  description?: string | null;
  priority: CreateTaskRequestPriority;
  position: number;
}

export type UpdateTaskRequestPriority =
  (typeof UpdateTaskRequestPriority)[keyof typeof UpdateTaskRequestPriority];

export const UpdateTaskRequestPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface UpdateTaskRequest {
  columnId?: number;
  title?: string;
  description?: string | null;
  priority?: UpdateTaskRequestPriority;
  position?: number;
}

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type GetTasksParams = {
  columnId?: number;
};