# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nature Kanban Task Manager API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: columns
    description: Kanban columns
  - name: tasks
    description: Kanban tasks
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      operationId: getCurrentAuthUser
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Auth status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"

  /login:
    get:
      operationId: beginBrowserLogin
      tags: [auth]
      summary: Start the OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to OIDC provider

  /callback:
    get:
      operationId: handleBrowserLoginCallback
      tags: [auth]
      summary: Complete the OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect after login

  /logout:
    get:
      operationId: browserLogout
      tags: [auth]
      summary: Log out
      responses:
        "302":
          description: Redirect after logout

  /mobile-auth/token-exchange:
    post:
      operationId: exchangeMobileAuthorizationCode
      tags: [auth]
      summary: Mobile token exchange
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeBody"
      responses:
        "200":
          description: Session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeResponse"

  /mobile-auth/logout:
    post:
      operationId: logoutMobileSession
      tags: [auth]
      summary: Mobile logout
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutMobileSessionResponse"

  /columns:
    get:
      operationId: getColumns
      tags: [columns]
      summary: Get all columns
      responses:
        "200":
          description: List of columns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Column"
    post:
      operationId: createColumn
      tags: [columns]
      summary: Create a column (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateColumnRequest"
      responses:
        "201":
          description: Created column
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Column"

  /columns/{id}:
    put:
      operationId: updateColumn
      tags: [columns]
      summary: Update a column (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateColumnRequest"
      responses:
        "200":
          description: Updated column
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Column"
    delete:
      operationId: deleteColumn
      tags: [columns]
      summary: Delete a column (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: Get tasks
      parameters:
        - name: columnId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [admin, user]
      required: [id, role]

    AuthUserEnvelope:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
          nullable: true
      required: [user]

    ExchangeMobileAuthorizationCodeBody:
      type: object
      properties:
        code:
          type: string
        state:
          type: string
        nonce:
          type: string
          nullable: true
        code_verifier:
          type: string
        redirect_uri:
          type: string
      required: [code, state, code_verifier, redirect_uri]

    ExchangeMobileAuthorizationCodeResponse:
      type: object
      properties:
        token:
          type: string
      required: [token]

    LogoutMobileSessionResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    Column:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, position, createdAt]

    CreateColumnRequest:
      type: object
      properties:
        title:
          type: string
        position:
          type: integer
      required: [title, position]

    UpdateColumnRequest:
      type: object
      properties:
        title:
          type: string
        position:
          type: integer

    Task:
      type: object
      properties:
        id:
          type: integer
        columnId:
          type: integer
        userId:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: string
          enum: [low, medium, high]
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, columnId, title, priority, position, createdAt]

    CreateTaskRequest:
      type: object
      properties:
        columnId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: string
          enum: [low, medium, high]
        position:
          type: integer
      required: [columnId, title, priority, position]

    UpdateTaskRequest:
      type: object
      properties:
        columnId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: string
          enum: [low, medium, high]
        position:
          type: integer