# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nature Kanban Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get the currently authenticated user
 */
export const GetCurrentAuthUserResponse = zod.object({
  user: zod
    .object({
      id: zod.string(),
      email: zod.string().nullish(),
      firstName: zod.string().nullish(),
      lastName: zod.string().nullish(),
      profileImageUrl: zod.string().nullish(),
      role: zod.enum(["admin", "user"]),
    })
    .nullable(),
});

/**
 * @summary Start the OIDC login flow
 */
export const BeginBrowserLoginQueryParams = zod.object({
  returnTo: zod.coerce.string().optional(),
});

/**
 * @summary Complete the OIDC login flow
 */
export const HandleBrowserLoginCallbackQueryParams = zod.object({
  code: zod.coerce.string().optional(),
  state: zod.coerce.string().optional(),
  iss: zod.coerce.string().optional(),
});

/**
 * @summary Mobile token exchange
 */
export const ExchangeMobileAuthorizationCodeBody = zod.object({
  code: zod.string(),
  state: zod.string(),
  nonce: zod.string().nullish(),
  code_verifier: zod.string(),
  redirect_uri: zod.string(),
});

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  token: zod.string(),
});

/**
 * @summary Mobile logout
 */
export const LogoutMobileSessionResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get all columns
 */
export const GetColumnsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  position: zod.number(),
  createdAt: zod.date(),
});
export const GetColumnsResponse = zod.array(GetColumnsResponseItem);

/**
 * @summary Create a column (admin only)
 */
export const CreateColumnBody = zod.object({
  title: zod.string(),
  position: zod.number(),
});

/**
 * @summary Update a column (admin only)
 */
export const UpdateColumnParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateColumnBody = zod.object({
  title: zod.string().optional(),
  position: zod.number().optional(),
});

export const UpdateColumnResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  position: zod.number(),
  createdAt: zod.date(),
});

/**
 * @summary Delete a column (admin only)
 */
export const DeleteColumnParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get tasks
 */
export const GetTasksQueryParams = zod.object({
  columnId: zod.coerce.number().optional(),
});

export const GetTasksResponseItem = zod.object({
  id: zod.number(),
  columnId: zod.number(),
  userId: zod.string().nullish(),
  title: zod.string(),
  description: zod.string().nullish(),
  priority: zod.enum(["low", "medium", "high"]),
  position: zod.number(),
  createdAt: zod.date(),
});
export const GetTasksResponse = zod.array(GetTasksResponseItem);

/**
 * @summary Create a task
 */
export const CreateTaskBody = zod.object({
  columnId: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  priority: zod.enum(["low", "medium", "high"]),
  position: zod.number(),
});

/**
 * @summary Update a task
 */
export const UpdateTaskParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTaskBody = zod.object({
  columnId: zod.number().optional(),
  title: zod.string().optional(),
  description: zod.string().nullish(),
  priority: zod.enum(["low", "medium", "high"]).optional(),
  position: zod.number().optional(),
});

export const UpdateTaskResponse = zod.object({
  id: zod.number(),
  columnId: zod.number(),
  userId: zod.string().nullish(),
  title: zod.string(),
  description: zod.string().nullish(),
  priority: zod.enum(["low", "medium", "high"]),
  position: zod.number(),
  createdAt: zod.date(),
});

/**
 * @summary Delete a task
 */
export const DeleteTaskParams = zod.object({
  id: zod.coerce.number(),
});