# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nature Kanban Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from "./authUserRole";

export interface AuthUser {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
  role: AuthUserRole;
}