# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nature Kanban Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from "./taskPriority";

export interface Task {
  id: number;
  columnId: number;
  userId?: string | null;
  title: string;
  description?: string | null;
  priority: TaskPriority;
  position: number;
  createdAt: Date;
}